【版本】2.5.12【介绍】优化了IOC【功能】1、自动完成Activity、Fragment、Dialog中View的绑定 2、MVP的支持 3、企业级日志系统 4、TextUtils、NumberUtils、IoUtils、ImageUtils等工具类 5、ORM数据库关系映射框架 6、多种颜色值代码 7、万能AdapterView适配器 8、ShadeView滑动颜色渐变控件 9、水平Tab滚动条 10、多行单选按钮组 11、字母导航控件 12、可嵌套在ScrollView内部的ListView 13、雷达扫描控件 14、流式布局 15、带清除按钮的文本框 16、高亮背景可游走的TextView 17、自动切换焦点的文本框和文本框组 18、动画框架 19、底部导航栏 20、仪表控件 21、标题栏 22、颜色拾取控件 23、动画框架2 24、侧滑菜单 25、带进度的按钮 26、GIF动画引擎 27、环形变速进度条 28、跑马灯 29、带翻转效果的ImageView 30、吞没事件的RelativeLayout 31、变色的TextView 32、圆形头像ImageView 33、侧滑菜单2 34、刮刮卡控件 35、幸运转盘 36、下拉刷新和上拉加载 37、基础对话框

+++ jackknife-util/src/main/java/com/lwh/jackknife/util/DensityUtils.java
@@ -24,13 +24,13 @@ public class DensityUtils {
     private DensityUtils() {
     }
 
-    public static int dp2px(float dpVal, Context context) {
-        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP,
+    public static float dp2px(float dpVal, Context context) {
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP,
                 dpVal, context.getResources().getDisplayMetrics());
     }
 
-    public static int sp2px(float spVal, Context context) {
-        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
+    public static float sp2px(float spVal, Context context) {
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
                 spVal, context.getResources().getDisplayMetrics());
     }
